requirements_essential.txt
# FisioFlow Backend - Dependências Essenciais

# Framework Web
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.5.3
Flask-CORS==4.0.0

# Banco de Dados
psycopg2-binary==2.9.7
SQLAlchemy==2.0.21

# Cache e Sessões
redis==5.0.0

# Autenticação e Segurança
PyJWT==2.8.0
bcrypt==4.0.1
passlib==1.7.4

# Validação e Serialização
marshmallow==3.20.1
Flask-Marshmallow==0.15.0

# APIs e Integrações
requests==2.31.0

# Configuração e Ambiente
python-dotenv==1.0.0

# Logging
structlog==23.1.0

# Utilitários de Data e Tempo
python-dateutil==2.8.2

# Processamento de Dados
pandas==2.1.1
numpy==1.25.2

# Testes
pytest==7.4.2
pytest-flask==1.2.0

# Servidor WSGI
gunicorn==21.2.0

